- Add MusicAlbum class - Add Initializer method - Add can_be_archived method
